Add types for terraform route
